requirements/staging.txt
-r base.txt
django-debug-toolbar==1.2.1
django-extensions==1.3.2